Правки по КР: - Исправлены констрейты у поля "Статус" для маленьких экранов - Удален неиспользуемый код прошлого хедера - Для картинок в посте отменены констрейнты - Ширина отступа после таблицы задана 0 - Выровнены отступы через cmd + a, cmd + i - Исправлены var на lat у атрибутов структуры
