+++ Navigation/FeedViewController.swift
@@ -17,7 +17,7 @@ class FeedViewController: UIViewController {
         button.addTarget(self, action: #selector(self.showPost), for: .touchUpInside)
         return button
     }()
-
+    
     private lazy var button2: UIButton = {
         let button = UIButton(frame: CGRect(x: 0, y: 0, width: 140, height: 35))
         button.backgroundColor = .systemTeal
@@ -26,7 +26,7 @@ class FeedViewController: UIViewController {
         button.addTarget(self, action: #selector(self.showPost), for: .touchUpInside)
         return button
     }()
-
+    
     private lazy var vStack: UIStackView = {
         let vStack = UIStackView(frame: CGRect(x: 100, y: 100, width: 140, height: 100))
         vStack.layer.borderColor = CGColor(red: 0, green: 0, blue: 0, alpha: 1)
@@ -44,7 +44,7 @@ class FeedViewController: UIViewController {
         self.view.backgroundColor = .white
         self.view.addSubview(self.vStack)
         self.vStack.center = self.view.center
-
+        
         // настраиваем вью, если темная тема
         isCurrentThemeDark()
     }
@@ -54,7 +54,7 @@ class FeedViewController: UIViewController {
         self.navigationController?.pushViewController(postScene, animated: true)
         postScene.title = post.title
     }
-
+    
     func isCurrentThemeDark(){
         if (traitCollection.userInterfaceStyle == UIUserInterfaceStyle.dark) {
             print("Current Theme is dark")
