+++ Navigation/Profile/LogInViewController.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 class LogInViewController: UIViewController {
-
+    
     private lazy var logoImage: UIImageView = {
         var image = UIImageView()
         image.image = UIImage(named: Constants.logoName)
@@ -19,7 +19,7 @@ class LogInViewController: UIViewController {
         image.translatesAutoresizingMaskIntoConstraints = false
         return image
     }()
-
+    
     private lazy var stackLogoImage: UIStackView = {
         let lStack = UIStackView()
         lStack.layer.borderColor = UIColor.white.cgColor
@@ -34,7 +34,7 @@ class LogInViewController: UIViewController {
         lStack.addArrangedSubview(logoImage)
         return lStack
     }()
-
+    
     private lazy var emailTextField: UITextField = {
         let email = UITextField()
         email.layer.borderColor = UIColor.lightGray.cgColor
@@ -47,7 +47,7 @@ class LogInViewController: UIViewController {
         email.layer.sublayerTransform = CATransform3DMakeTranslation(16, 0, 0)
         return email
     }()
-
+    
     private lazy var passwordTextField: UITextField = {
         let password = UITextField()
         password.layer.borderColor = UIColor.lightGray.cgColor
@@ -62,7 +62,7 @@ class LogInViewController: UIViewController {
         password.layer.sublayerTransform = CATransform3DMakeTranslation(16, 0, 0)
         return password
     }()
-
+    
     private lazy var stackTextFields: UIStackView = {
         let tStack = UIStackView()
         tStack.layer.borderColor = UIColor.lightGray.cgColor
@@ -78,7 +78,7 @@ class LogInViewController: UIViewController {
         tStack.addArrangedSubview(passwordTextField)
         return tStack
     }()
-
+    
     private lazy var logInButton: UIButton = {
         let button = UIButton()
         button.tag = Constants.logInButtonTap
@@ -90,7 +90,7 @@ class LogInViewController: UIViewController {
         button.addTarget(self, action: #selector(goToProfileViewController), for: .touchUpInside)
         return button
     }()
-
+    
     private lazy var stackButton: UIStackView = {
         let bStack = UIStackView()
         bStack.distribution = .fillEqually
@@ -100,19 +100,19 @@ class LogInViewController: UIViewController {
         bStack.addArrangedSubview(logInButton)
         return bStack
     }()
-
+    
     private lazy var scrollView: UIScrollView = {
         let scroll = UIScrollView()
         scroll.translatesAutoresizingMaskIntoConstraints = false
         return scroll
     }()
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         setupView()
         setupGesture()
     }
-
+    
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         NotificationCenter.default.addObserver(self,
@@ -124,13 +124,13 @@ class LogInViewController: UIViewController {
                                                name: UIResponder.keyboardDidHideNotification,
                                                object: nil)
     }
-
+    
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         // чтобы клавиатура сразу была в поле email
-      //  emailTextField.becomeFirstResponder()
+        //  emailTextField.becomeFirstResponder()
     }
-
+    
     private func setupView(){
         self.view.backgroundColor = .white
         navigationController?.navigationBar.isHidden = true
@@ -140,30 +140,30 @@ class LogInViewController: UIViewController {
         scrollView.addSubview(stackLogoImage)
         scrollView.addSubview(stackTextFields)
         scrollView.addSubview(stackButton)
-
+        
         NSLayoutConstraint.activate([
             scrollView.topAnchor.constraint(equalTo: view.topAnchor),
             scrollView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
             scrollView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
             scrollView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-
+            
             stackLogoImage.topAnchor.constraint(equalTo: scrollView.topAnchor, constant: 120),
             stackLogoImage.heightAnchor.constraint(equalToConstant: 100),
             stackLogoImage.widthAnchor.constraint(equalToConstant: 100),
             stackLogoImage.bottomAnchor.constraint(equalTo: stackTextFields.topAnchor, constant: -120),
             stackLogoImage.centerXAnchor.constraint(equalTo: scrollView.centerXAnchor),
-
+            
             stackTextFields.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
             stackTextFields.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
             stackTextFields.heightAnchor.constraint(equalToConstant: 100),
-
+            
             stackButton.topAnchor.constraint(equalTo: stackTextFields.bottomAnchor, constant: 16),
             stackButton.leadingAnchor.constraint(equalTo: stackTextFields.leadingAnchor),
             stackButton.trailingAnchor.constraint(equalTo: stackTextFields.trailingAnchor),
             stackButton.heightAnchor.constraint(equalToConstant: 50)
         ])
     }
-
+    
     @objc private func goToProfileViewController(sender: UIButton){
         if sender.tag == Constants.logInButtonTap {
             hideKeyboard()
@@ -172,29 +172,29 @@ class LogInViewController: UIViewController {
             self.present(goToProfileViewController, animated: true)
         }
     }
-
+    
     private func setupGesture(){
         let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboard))
         view.addGestureRecognizer(tapGesture)
     }
-
+    
     @objc func hideKeyboard(){
         view.endEditing(true)
         scrollView.setContentOffset(.zero, animated: true)
     }
-
+    
     @objc private func didShowKeyboard(_ notification: Notification){
         if let keyboardFrame = notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue {
             let keyboardRectangle = keyboardFrame.cgRectValue
             let keyboardHeight = keyboardRectangle.height
-
+            
             let logInButtomBottomPointY =
             stackButton.frame.origin.y +
             logInButton.frame.origin.y +
             logInButton.frame.height
-
+            
             let keyboardOriginY = view.frame.height - keyboardHeight
-
+            
             var offSet: CGFloat = CGFloat()
             if keyboardOriginY <= logInButtomBottomPointY {
                 offSet = logInButtomBottomPointY - keyboardOriginY + 16
@@ -202,7 +202,7 @@ class LogInViewController: UIViewController {
             scrollView.contentOffset = CGPoint(x: 0, y: offSet)
         }
     }
-
+    
     @objc private func didHideKeyBoard(_ notification: Notification){
         hideKeyboard()
     }
