+++ Navigation/Profile/PostTableViewCell.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 class PostTableViewCell: UITableViewCell {
-
+    
     private lazy var titlePostLabel: UILabel = {
         let title = UILabel()
         title.font = .boldSystemFont(ofSize: 20)
@@ -17,7 +17,7 @@ class PostTableViewCell: UITableViewCell {
         title.translatesAutoresizingMaskIntoConstraints = false
         return title
     }()
-
+    
     private lazy var imagePostView: UIImageView = {
         let image = UIImageView()
         image.contentMode = .scaleAspectFit
@@ -26,7 +26,7 @@ class PostTableViewCell: UITableViewCell {
         image.translatesAutoresizingMaskIntoConstraints = false
         return image
     }()
-
+    
     private lazy var descriptionPostLabel: UILabel = {
         let description = UILabel()
         description.font = .systemFont(ofSize: 14)
@@ -35,7 +35,7 @@ class PostTableViewCell: UITableViewCell {
         description.translatesAutoresizingMaskIntoConstraints = false
         return description
     }()
-
+    
     private lazy var likePostLabel: UILabel = {
         let like = UILabel()
         like.font = .systemFont(ofSize: 16)
@@ -43,7 +43,7 @@ class PostTableViewCell: UITableViewCell {
         like.translatesAutoresizingMaskIntoConstraints = false
         return like
     }()
-
+    
     private lazy var viewPostLabel: UILabel = {
         let view = UILabel()
         view.font = .systemFont(ofSize: 16)
@@ -51,7 +51,7 @@ class PostTableViewCell: UITableViewCell {
         view.translatesAutoresizingMaskIntoConstraints = false
         return view
     }()
-
+    
     private lazy var authorPostLabel: UILabel = {
         let author = UILabel()
         author.font = .systemFont(ofSize: 16)
@@ -59,16 +59,16 @@ class PostTableViewCell: UITableViewCell {
         author.translatesAutoresizingMaskIntoConstraints = false
         return author
     }()
-
+    
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         self.setupView()
     }
-
+    
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
+    
     func setup(for model: Post){
         titlePostLabel.text = model.title
         authorPostLabel.text = model.author
@@ -77,7 +77,7 @@ class PostTableViewCell: UITableViewCell {
         likePostLabel.text = "Likes: \(model.likes)"
         viewPostLabel.text = "Views: \(model.views)"
     }
-
+    
     func setupView(){
         contentView.addSubview(titlePostLabel)
         //contentView.addSubview(authorPostLabel)
@@ -85,26 +85,26 @@ class PostTableViewCell: UITableViewCell {
         contentView.addSubview(imagePostView)
         contentView.addSubview(likePostLabel)
         contentView.addSubview(viewPostLabel)
-
+        
         NSLayoutConstraint.activate([
             titlePostLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
             titlePostLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-
+            
             imagePostView.topAnchor.constraint(equalTo: titlePostLabel.bottomAnchor, constant: 8),
             imagePostView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
             imagePostView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
-            imagePostView.widthAnchor.constraint(equalToConstant: UIScreen.main.bounds.width),
-            imagePostView.heightAnchor.constraint(equalTo: imagePostView.widthAnchor, multiplier: 1.0),
-
+            //imagePostView.widthAnchor.constraint(equalToConstant: UIScreen.main.bounds.width),
+            //imagePostView.heightAnchor.constraint(equalTo: imagePostView.widthAnchor, multiplier: 1.0),
+            
             descriptionPostLabel.topAnchor.constraint(greaterThanOrEqualTo: imagePostView.bottomAnchor, constant: 8),
             descriptionPostLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
             descriptionPostLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
             descriptionPostLabel.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -26),
-
+            
             likePostLabel.topAnchor.constraint(greaterThanOrEqualTo: descriptionPostLabel.bottomAnchor, constant: -8),
             likePostLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
             likePostLabel.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
-
+            
             viewPostLabel.topAnchor.constraint(greaterThanOrEqualTo: descriptionPostLabel.bottomAnchor, constant: -8),
             viewPostLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
             viewPostLabel.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)
